httpx/httpx.go netx.go: dont wrap *tls.Conn
There is code in net/http that casts a net.Conn to *tls.Conn to
inspect the ALPN for upgrading to http2.

As such, wrapping a *tls.Conn is a bad idea.

Change the code to always use an ExternalConnID derived from
the five tuple to cross reference measurements.

Keep wrapping the net.Conn returned by cleartext dialing, since
this seems to work okay in my tests.