httpx: Implement http2
After the previous patch, which paved the way, the only remaining
bit is to use a *tls.Config with the proper ALPN config.

In addition to that, we need to mess with the transport internals.

All of this can be done by just using the proper config function.

I would rather not use the built-in h2 support because that is
going to change the tls.Config in a more implicit way.