Changed my mind: let's return an interface for the resolver
When we can return an interface of the stdlib, like RoundTripper, we
do that because this is more extensible.

However, already for the Dialer we augment the stdlib. So it makes
sense here to return an interface because this allows us to use the
current approach or an approach where we use, say, miekg/dns.