dialerbase: pass handler as DialHostPort argument
We're now starting to do some serious business. Our objective is to
decouple the handler that is handling a (chain of) request(s) from
the specific httpx.Transport. This means that the handler to be used
at any given time must depend somehow on the context [*].

So, let's start at the very lowest level, dialerbase, to introduce
the concept of the handler currently being used, as opposed to the
handler that has been set in the constructor.

It will be easy for the callers of dialerbase to adapt but we will
start our journey in handler decoupling.

[*] I meant context, not the Context object in the context package!