Introduce dialercontext as a simpler dialerapi
The plan is to use dialercontext in http code, such that we can use the
context to pass around the measurement handler to use.

This is easy to do, because DialContext receives a context as its first
argument. Thus, we could attach the handler to such context and then
obtain the handler from such context and use the handler.

However, to move in this direction we need to clearly split dialerapi,
which also includes TLS APIs that do not receive a context, e.g.,
`DialTLS`, from dialercontext, which only implements `DialContext`.

After this split, `httpx` code will use dialercontext and perform its
own TLS. We will read TLS events using httptrace. Instead, `next` code
can stay with dialerapi, since we don't have the issue of attributing
events to specific round-trips for `netx` code. This means we will
use slighly different code for TLS depending on whether we're using
`httpx` or `netx`. That said, we can rather easily overcome this issue
by factoring common code for initializing `*tls.Config`, if needed.

The code has been copied and adapted in the most trivial way to get
as a result a smaller implementation.

At this stage, it is not yet clear to me what to do with dialerapi. It
seems plausible that later I'll reimplement it as a wrapper around of
dialercontext. So, I've added a comment to signal that there is some code
duplication, and some more pending refactoring.