Add code to bind handler to context
Now that we've started to create a framework for passing around a
context with attached handler, we need to make it possible to really
attach an handler to a context.

This diff main concern is to make that possible. Also, because in
many cases we will have a local handler, obtained from the context,
plus a global handler, set in the constructor, also implement the
facilities to create the composition of two handlers.

This way, we can easily emit once and fan-out to many handlers.