httptransport: allow saving to multiple handlers
We finally reached the point where we see some light and we are
able to store a bunch of events that pertain to a round trip into
a specific place that is not a catch-all handler.

The core idea is that, during the round trip, we will emit to a
handler that is the composition of the one passed to the constructor
of httpx.Transport (likely a logger) and the one that has possibly
been set inside the current context (likely a measurement saver).

After this change, there is a bunch of events related to the HTTP
round trip that are emitted also on the context-bound handler.

However, we are missing many more events. What it remains to do now
is to further refactor the code such that all events that we are
used to see in the global handler could also appear in the context one.