httpx: replace dialerapi with dialercontext
As started in precedence, dialerapi is not good for us because it
encourages you to DialTLS, which has no context.

To finally be able to see network-level events from httpx again, we
actually need to switch over to dialercontext, which encourages a
person to just DialContext. To make sure TLS is working again, then
we need to apply some changes in the way in which the HTTP transport
interacts with the dialercontext. Mainly, we will unset enough of
the fields to nudge the transport to perform its own TLS handshake
using as DialContex the dialercontext.Dialer method.

Because dialerapi (and dnsconf) are quite complex and the code is
perhaps too much coupled, as side effect of this change is that we
have broken ConfigureDNS for `httpx`. It will not be immediate to
fix this; on the contrary, it will take several commits. The way in
which I'm dealing with this is that I changed the code to cheat to
the tests pretending ConfigureDNS is still working. I think this is
fine because all these commits will be merged together.

On the flip side, with the implementation of this change, we now
see again all the network level events.