httptransport: export TLS handshake event
This commit improves the situation by also exporting the TLS handshake
to the current handler(s). Because we are setting a httptrace tracer
into a context private of the current round-trip request, there should
only be a single startHandshake event followed by a single stopHandsake.

After this commit, the only remaining part to be addressed is to
make sure we see again custom-DNS events. This will be our Ardennes.