oodns: write context aware doh transport
Now we can capture events occurring during a round trip using
a doh transport using a context.

This has been implemented only for oodns, because it really
seems that godns is too much of a hack to bother.

After this change, we can also measure DoH events as before
but they are rooted on the context.

Also: note how the new code is much more independent of netx
and httpx than the previous implementation.