oodns: rewrite the netgo transport
After this change, the `netgo` transport works as before, but only
on platforms where netgo is supported (e.g. Unix). And of course the
DNS events are now context-rooted.

While there, notice that the name used in cmd/httpclient/main.go was
not consistent with the spec and make it consistent.

Also: have tried to make this code more general and independent
of specific implementation details of netx.