httpx: refactor ConfigureDNS and improve it
This diff refactores the new implementation of httpx's ConfigureDNS to
become an independent package returning a dnsx.Client.

This is similar in concept to dnsconf. It is basically the dnsconf
equivalent for dialerfull/dialercontext. So, I've also marker the
dnsconf package as deprecated and to be replaced.

After this diff, we've completely repaired ConfigureDNS and just a
little bit of yak shaving separates us from declaring success.