DESIGN.md: context- or request-rooted measurements
We still want a global handler, as that is usefulf for logging.

But we also want a context or a request to be the "root" and all the
related measurements to be associated with that root.

This has been implemented in ooni/probe-engine by basically forcing
no parallelism, however this is potentially limiting.

Instead, we should implement it here in a way that does not prevent
us from running parallel measurements.

What's more, when looking into what a specific request did, we also
want to know the moment when an HTTP round trip starts.

This is actually easy to do, as it suffices to define and emit a
new, very minimal event to mark that moment.

I believe this enhancement would provide us with enough clarity and
flexibility that we won't need to implementing joining the measurements
in Go by using the five tuple as key.

To implement all of this, we need lots of yak shaving and changes in
the codebase. This is just the first commit, where we declare what we
want to do. Subseqent commits will try to do that.