dnsovertcp: decouple from netx/httpx
Same purpose of #28 and #27. We want to reach a situation where we can
construct a DNS transport that does not depend on netx/httpx.

The general objective here is (1) to decouple unrelated parts of the
tree and (2) to use dependency injection to enable writing wrappers for
any piece of code, so to enable follow-up measurements when any step
we're performing seems to return something fishy.

While working on this diff, I noticed that there was inconsistency
in the way in which we manage DNS:

- some resolvers tolerated domain input, others did not

- some resolvers automatically added ports in some cases

I've therefore made sure all resolvers are constructed in a way that
is robust to supplying DNS names as well as to missing ports.