refactor(dialerapi): decouple from other packages
Merge dialerbase into dialerapi since dialerbase was basically
just a small dialerapi implementation detail.

Ensure that the input parameters to the constructor and the
dependencies of dialerbase are not weird stuff.

Try to limit exposure of implementation details or weird types
as much as possible. Start thinking about deprecating some code
parts that are not much used and tricky.