Replace godns with oodns
The godns was bringing with it a significant baggage of complexity
however it did not work on all platforms.

As such, it has become an obvious target for pruning. It is in the
way of more refactoring and it's not adding value.

While there, also fix a bunch of small issues that I've met.