Allow composing dnsclients together
This starts to show the direction where I'm heading:

1. use interfaces more often

2. compose behaviours by composing interfaces

3. use context rather than global handlers

This is exemplified by delegating emitting the Resolve event to a
wrapper of a generic dnsclient, which may even be system.

In a follow-up PR, I'll add extra code for checking whether an IP
address is legit and possibly retry if not.

While there, move the basic interfaces that qualify DNS into model
because I constantly clash with them being in dsnx when refactoring,
since their presence and their low level usage basically makes the
dnsx package useless, as I can't add functions there.