netx: move beginning and handler at toplevel
All the netx code now does not depend on a preconfigured
beginning and handler, except toplevel code.

This is very helpful because it allows us to specify different
contexts depending on circumstances.

Yet, the public API has not changed because there is still
a notion of the binding of a dialer with a handler.

What I like about this change is that it allows us to evolve
our code potentially in two directions:

1. we keep a single handler and we provide correct information
concerning the current TransactionID

2. we switch to a rooted handlers model where the handler is
passed as part of the context

Now, internals in netx already follow model #2, while the toplevel
API is still using model #1.

The part that needs refactoring now is httpx.