httpx: move beginning and handler to toplevel
Like for 6dfb9d02d0c65237e520ced6615b55297a0cc447, make the
implementation of httpx independent of a specfic handler.

Just set the handler configured in httpx.Transport when we're
about to start a round trip, so we track events.

Now we can finally fix the bug that it is difficult to
process data emitted by this library.