[DATA CHANGE] Clearly identify the round trip
The round trip is what happens since we start thinking about
sending a request until when we receive response.

This is a clearly defined operation with a well defined result
indicating error or HTTP response headers.

As such, emit its events in the proper model, and make sure
that we also emit the error that occurred (if any).

This change simplifies data analysis.

While there, stop duplicating docs in the design document, since
we are going to change events more times in the future.