Nervous system: nervous resolver that handles bogons
I am not 100% sold yet on what is the actionable top-level event
to in order to clarify in the log we've seen a bogon.

Another good actionable idea could be to flag the presence of
bogons (maybe along with DialID?) in MeasurementRoot, but this
is also something I'd like to think about.

This code in here is good for me to start understand it and then
applying further changes as I get more insights.