nervousresolver: fix processing of bogons
For now let us limit to act if we see any bogon. If we see at
least one, we cannot trust the upstream DNS. If we don't see any,
or we see an error, for now don't use the secondary resolver as
this complicates making other techniques possible.

I believe we may reconsider the latter choice when we have
some more experience with this part of the tree and we have
run more testing with jafar and/or in the field.