Make sure our DNS retry policy is sane
Implement retries for queries. Tested with unit testing as well
as configuring ooni/jafar to drop traffic to 9.9.9.9.

No need for fancy retry, after all.

No need to act at TCP level because the kernel already retries.

We should be good. The other level of retry is the one that
probably concerns individual URL measurements, not to be done here.