Rework nervous resolver and make it opt-in
I realised it was not a good idea to have it hidden and enabled
by default, as it prevented other kind of testing.

So, allow using it as opt-in.

Also, while there, make it better by using it whenever we fail
to query the DNS so to have more DNS robustness.

I think with this I'm near the bottom of my netx TODO list.