Adapt the HTTP configuration more to OONI needs
We need to observe real headers, so reduce the magic performed
under the hood by Go. The really good thing would be to be able
to store information about the round trip more easily.

Events are great for debugging, but some higher level abstraction
that takes into account all oddities is probably required.