Add enough porcelain to make Telegram simple
I've determined that it's better to add more porcelain here as
an extension, and keep ooni/probe-engine simpler.

Because, if something breaks in the data format, we want to catch
it inside here rather than cathing it in the other repo.

So, ooni/probe-engine will use the porcelain that I've just
added to run telegram and will process results, that are now
in a format by which processing is very easy.

Basically, ooni/probe-engine will just need to filter out some
bits of information it doesn't need and call .Error() for errors.

Note that the porcelain I'm adding here is also enough to
implement web_connectivity's core and possible other experiments.

Reference issue: https://github.com/ooni/probe-engine/issues/86