nervousresolver: use cloudflare IP address
We should probably introduce a longer chain of fallbacks in the
nervous resolver, but, for now, using the IP address is good enough
because it can protect us against the whole DNS being down.

This happened to me in Italy, when "Vodafone rete sicura" had
several hours of outage and the default DNS was not working as
intended, while there was IP connectivity.

Reference issue: https://github.com/ooni/probe-engine/issues/88