http: change body reading strategy
Always read snapshots of all bodies. Limit snapshot to 1<<17
that is good enough because:

1. for requests, we can always save them since we issue them but
it may be useful to save small requests

2. for responses, we can always save them except redirects and
DoH, and in both cases we are talking about small bodies

Fixes the output emitted by a DoH query.

We can now probably further simplify porcelain.go in a followup PR.