internal: use global doh client with NoHandler
As explained in the code, this allows us to avoid worrying about
closing the connections that an interim HTTP client may have
opened, saving lots of API complexity. Because setting NoHandler
as the handler is really the common case, this is a good enough
not-so-complex solution to avoid the technical debt caused by
failing to close the open connections of an *http.Client.