roundTripQuery: trust the first error more (#114)
The first error is the one that is most likely to be caused by the network. Subsequent errors are more likely to be caused by context deadlines. So, the first error is attached to an operation, while subsequent errors may possibly not be. If so, the resulting failing operation is not correct.