DESIGN.md: simplify and reorganize after recent changes (#119)
No real API change. What changes is the way in which I think
about next, so updated the document to reflect that.

While there, define a bunch of more tidy APIs where we can
avoid passing in handlers.NoHandler explicitly.

Also, the real replacement is the transport and not the client
which is just syntactic sugar around it. Make it clear.