dns: don't automatically fail when a bogon is found
In principle, we may want to measure a website from a place where it has
a bogon address, so we don't want to fail in such case.

At the same time, it may be we're giving too much importance on block
pages, and we should instead cut it short when we find a bogon.

Yet, when we're using OONI infrastructure, we certainly do not want to
continue, if we're returned a bogon by the resolver.

I believe we'll need to rediscuss this set of assumptions when the
design document with @fortuna makes more progress.

For now it is fine to not aggressively enable this bogon functionality
by default and to keep thinking about it.