Make sure regress tests are all green on travis (#145)
* ooniresolver_test.go: clarify when this test may fail

It certainly fails in my house. I have Vodafone Italy and I have not
disabled the "rete sicura" option, which does DNS hijacking.

I used to have it disabled, but then I moved and now I am again using
the default value of that setting.

So, I figured it made sense to run all tests and mark those which are
going to fail if you're under a similar setup.

(Making them pass under such conditions is probably not what I want.)

* porcelain_test.go: repair broken integration test

The test assumed there was no redirect when accessing ooni.io but it has
recently changed to redirect to ooni.org.

By changing the input URL for this integration test, we fix it.