http: last resort error wrapping must be in internal
We were previously wrapping the error in porcelain, which was not
100% correct, because technically porcelain is just a set of helpers
on the top the underlying functionality.

Instead, perform this kind of wrapping in the internal RoundTripper.

With also this change committed to master, we have severed the internal
ties between x/porcelain and the rest of netx. This is functional to
my plan of forking x/porcelain inside of probe-engine.